Change the time picking to take it as an UInt64

+++ tasks/Task.cpp
@@ -110,7 +110,7 @@ struct controldev_websocket::MessageDecoder{
 
     base::Time getTime(){
         return base::Time::fromMilliseconds(
-            jdata["time"].asDouble()
+            jdata["time"].asUInt64()
         );
     }
 };
